refactor: 날씨 데이터 정리 배치 Job/Step 정의
